% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data2.R
\name{generate_data2}
\alias{generate_data2}
\title{generate_data2}
\usage{
generate_data2(n_obs, corr, effect, ld, alp)
}
\arguments{
\item{n_obs}{number of observation}

\item{corr}{correlation of latent variable}

\item{effect}{effect of moderator}

\item{ld}{factor loading of one latent variable to its indicators}

\item{alp}{reliability of latent variable}
}
\value{
data stimulated from arguement setting
}
\description{
non-normal

generate data from Cheung et al.(2021). Note the reliability used here is omega
}
\examples{
n_obs = 100
corr = 0.1
effect = 0.12
ld = c(1,1,1,1)
alp = 0.9
generate_data(n_obs, corr, effect, ld, alp)
}
