% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defclass.R
\docType{class}
\name{Silp-class}
\alias{Silp-class}
\title{Define silp class}
\description{
Define silp class
}
\section{Slots}{

\describe{
\item{\code{raw_model}}{The user-specified \code{lavaan} syntax model.}

\item{\code{rapi_model}}{The revised model with the RAPI method.}

\item{\code{time}}{The operation time for \code{silp} (in seconds).}

\item{\code{npd}}{Logical. Whether the nearest positive definite matrix is used.}

\item{\code{raw_data}}{The input data.}

\item{\code{fa}}{An object of class \code{lavaan} representing the CFA result.}

\item{\code{reliability}}{The reliability index.}

\item{\code{composite_data}}{The composite data for RAPI.}

\item{\code{pa}}{The result of \code{silp}.}

\item{\code{boot}}{The results of \code{resilp} from R bootstrap samples.}

\item{\code{origine}}{The original \code{silp} estimation.}

\item{\code{time_resilp}}{The operation time for \code{resilp} (in seconds).}
}}

