% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silp.R
\name{silp}
\alias{silp}
\title{silp}
\usage{
silp(model, data, double = "double", reliability = "omega", type = "cov", ...)
}
\arguments{
\item{model}{a lavaan syntax model}

\item{data}{dataset for lavaan sem}

\item{double}{Single or Double. whether double mean center is used for product indicator. Default double}

\item{reliability}{which kind of reliability is used to estimate error variance If "alpha", Cronbach alpha reliability
is used. "omega" for omega reliability is used. Default omega}

\item{type}{which type of input is used in lavaan sem, default cov (covariance matrix) or data(raw data)}

\item{...}{other parameters pass to lavaan sem.}
}
\value{
raw data is the original data passed to arg data.
fa is the cfa model of lavaan class object.
reliability is the estimated reliability form data
new attribute include raw_model, rapi_model, silp_time
}
\description{
silp
}
\examples{
n_obs = 100
corr = 0.1
effect = 0.12
ld = c(1,1,1,1)
alp = 0.9
data = generate_data(n_obs, corr, effect, ld, alp)
model = "
  fy =~ y1 + y2 + y3 + y4
  fx =~ x1 + x2 + x3 + x4
  fz =~ z1 + z2 + z3 + z4
  fy ~  fx + fz + fx:fz
"
silp(model, data)
}
