% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(n_obs, corr, effect, ld, alp, effect_x = 0.4, effect_z = 0.2)
}
\arguments{
\item{n_obs}{number of observation}

\item{corr}{correlation of latent variable}

\item{effect}{effect of moderator}

\item{ld}{factor loading of one latent variable to its indicators}

\item{alp}{reliability of latent variable}

\item{effect_x}{direction effect of x}

\item{effect_z}{direction effect of z}
}
\value{
data stimulated from arguement setting
}
\description{
generate data from Cheung et al.(2021). Note the reliability used here is omega
}
\examples{
n_obs = 100
corr = 0.1
effect = 0.12
ld = c(1,1,1,1)
alp = 0.9
generate_data(n_obs, corr, effect, ld, alp)
}
