% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resilp.R
\name{resilp}
\alias{resilp}
\title{resilp}
\usage{
resilp(fit, R = 300, progress = T, sig_level = 0.05)
}
\arguments{
\item{fit}{result object from silp}

\item{R}{number of bootstrap. Default 300}

\item{progress}{whether there is a progress bar}

\item{sig_level}{significant level. Default 0.05}
}
\value{
parameter table of bootstrap outcome of silp
}
\description{
resilp
}
\examples{
n_obs = 100
corr = 0.1
effect = 0.12
ld = c(1,1,1,1)
alp = 0.9
data = generate_data(n_obs, corr, effect, ld, alp)
model = "
  fy =~ y1 + y2 + y3 + y4
  fx =~ x1 + x2 + x3 + x4
  fz =~ z1 + z2 + z3 + z4
  fy ~  fx + fz + fx:fz
"
fit = silp(model, data)
resilp(fit, R = 10)
}
