% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defclass.R
\docType{class}
\name{Silp-class}
\alias{Silp-class}
\title{Define silp class}
\description{
Define silp class
}
\section{Slots}{

\describe{
\item{\code{raw_model}}{character.}

\item{\code{rapi_model}}{character.}

\item{\code{time}}{numeric.}

\item{\code{type}}{character.}

\item{\code{raw_data}}{data.frame.}

\item{\code{fa}}{lavaan.}

\item{\code{reliability}}{data.frame.}

\item{\code{composite_data}}{data.frame.}

\item{\code{pa}}{lavaan.}

\item{\code{boot}}{data.frame.}

\item{\code{statistic}}{data.frame.}

\item{\code{origine}}{data.frame.}

\item{\code{time_resilp}}{numeric.}
}}

\section{Method}{
  \describe{
    \item{summary}{\code{signature(object = "Silp", method = Bootstrap)}: Returns
      the summary result of silp or resilp. Method is for resilp only. If method = Bootstrap, percentile bootstrap result is presented. If method = BC_b, biased-correct bootstrap result is presented. }
    }
}

